src/test/java/com/guru99/demo/TestCase/LoginPageTest.java
package com.guru99.demo.TestCase;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.guru99.demo.TestBase.TestBase;
import com.guru99.demo.TestPage.LoginPage;

public class LoginPageTest extends TestBase {

	LoginPage loginPage;

	public LoginPageTest() {
		super();
	}

	@BeforeMethod
	public void setUp() {
		initialization();
		loginPage = new LoginPage();
	}

	@Test
	public void verifyReset() {
		String data[] = loginPage.reset(userid, password);
		if (data[0].equals("") && data[1].equals("")) {
			Assert.assertTrue(true);
		} else {
			Assert.assertTrue(false);
		}
	}

	@Test
	public void verifyLogin() {
		loginPage.login();
		if (driver.getPageSource().contains(userid)) {
			Assert.assertTrue(true);
		} else {
			Assert.assertTrue(false);
		}
	}

	@AfterMethod
	public void tearDown() {
		driver.quit();
	}
}
